Antiassociative Algebra

Methods to deal with the free antiassociative algebra over the reals with an arbitrary number of indeterminates. Antiassociativity means that (xy)z = -x(yz). Antiassociative algebras are nilpotent with nilindex four (Remm, 2022, <doi:10.48550/arXiv.2202.10812>) and this drives the design and philosophy of the package. Methods are defined to create and manipulate arbitrary elements of the antiassociative algebra, and to extract and replace coefficients. A vignette is provided.

Details

The DESCRIPTION file: This package was not yet installed at build time.
Index: This package was not yet installed at build time.

Functionality to work with the free antiassociative algebra in R. The hex sticker features an image taken from Hoffnung (1959) in which musical concepts [pizzicato, crescendo, etc] are given whimsical visual form. The character on the hex sticker is captioned “A Discord”: Hoffnung's interpretation of the musical concept of dissonance. In the book, the preceding image was a “chord”, evoking harmony. The discord, on the other hand, embodies–for me at least–antiassociativity: everything is wrong, wrong, wrong.

Author

Robin K. S. Hankin [aut, cre] (ORCID: <https://orcid.org/0000-0001-5982-0415>)

Maintainer: Robin K. S. Hankin <hankin.robin@gmail.com>

References

Hoffnung G (1959). Hoffnung's Acoustics. Dobson.

See also

Examples

x <- raaa()
x
#> free antiassociative algebra element:
#> +3b +5d +2c.d +3d.d +1(b.a)b +3(b.d)d +1(c.a)d
y <- raaa()

x+y
#> free antiassociative algebra element:
#> +3a +6b +7d +1b.b +2b.c +2c.d +2d.c +3d.d +4(a.c)c +1(b.a)b +2(b.d)c +3(b.d)d
#> +1(c.a)d +2(d.b)c
x*y
#> free antiassociative algebra element:
#> +9b.a +9b.b +6b.d +15d.a +15d.b +10d.d -3(b.b)b -6(b.b)c -6(b.d)c +6(c.d)a
#> +6(c.d)b +4(c.d)d -5(d.b)b -10(d.b)c +9(d.d)a +9(d.d)b -10(d.d)c +6(d.d)d